compress and extract in ram, 1 channel color

+++ src/main/java/encode/Node.java
@@ -1,10 +1,10 @@
 package encode;
 
 public class Node {
-  public short value;
+  public int value;
   public Node next;
 
-  public Node(short value) {
+  public Node(int value) {
     this.value = value;
     this.next = null;
   }
