+++ src/main/java/utils/Print.java
@@ -3,7 +3,7 @@ package utils;
 import encode.Node;
 
 public class Print {
-  public static void print1DArray(String s, short[] str) {
+  public static void print1DArray(String s, int[] str) {
     System.out.println(s);
     for (int i = 0; i < str.length; i++) {
       System.out.print(str[i]);
@@ -12,7 +12,7 @@ public class Print {
     System.out.println();
   }
 
-  public static void print2DArray(String s, short[][] str) {
+  public static void print2DArray(String s, int[][] str) {
     System.out.println(s);
     int m = str.length;
     int n = str[0].length;
